+++ auth-db/queries/email-auth.test.ts
@@ -12,42 +12,6 @@ describe("email-auth queries", () => {
     await db.users.deleteAll();
   });
 
-  describe("verifyEmail", () => {
-    it("should verify email and clear verification token", async () => {
-      const user = await db.users.create({
-        email: "test@example.com",
-        createdAt: new Date(),
-      });
-
-      const passwordHash = Buffer.from([1, 2, 3]);
-      const token = "verification-token";
-      const now = new Date();
-      now.setMilliseconds(0);
-      const expiresAt = new Date(now.getTime() + 15 * 60 * 1000); // 15 minutes from now
-
-      await db.emailAuth.create({
-        userId: user.id,
-        email: user.email,
-        passwordHash,
-        verificationToken: token,
-        verificationTokenExpiresAt: expiresAt,
-      });
-
-      const updated = await db.emailAuth.verifyEmail(user.id);
-      expect(updated).toMatchObject({
-        verifiedAt: expect.any(Date),
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-      });
-    });
-
-    it("should throw EmailAuthNotFoundError when user not found", async () => {
-      await expect(db.emailAuth.verifyEmail(999)).rejects.toThrow(
-        EmailAuthNotFoundError,
-      );
-    });
-  });
-
   describe("getByVerificationToken", () => {
     it("should get email auth by verification token", async () => {
       const user = await db.users.create({
