+++ auth-db/queries/email-auth/index.ts
@@ -6,3 +6,4 @@ export * from "./update-password-hash.ts";
 export * from "./get-by-forgot-password-token.ts";
 export * from "./update-password.ts";
 export * from "./clear-forgot-password-token.ts";
+export * from "./verify-email.ts";
